# face_recognition_system/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Face Recognition System API",
    "description": "\n        ## 人脸识别系统 API\n\n        这是一个完整的人脸识别和门禁控制系统的RESTful API。\n\n        ### 主要功能\n        - **人员管理**: 添加、更新、删除和搜索人员信息\n        - **人脸特征管理**: 上传人脸图像，提取和管理人脸特征\n        - **实时人脸识别**: 识别和验证人脸身份\n        - **门禁控制**: 基于人脸识别的访问控制\n        - **统计报告**: 访问日志分析和统计报告\n        - **用户管理**: 系统用户和权限管理\n        - **系统管理**: 系统配置、备份和监控\n\n        ### 认证方式\n        API使用JWT (JSON Web Token) 进行认证。获取令牌后，在请求头中包含：\n        ```\n        Authorization: Bearer <your_access_token>\n        ```\n\n        ### 权限系统\n        - **read**: 读取权限 - 可以查看数据\n        - **write**: 写入权限 - 可以创建和修改数据\n        - **admin**: 管理员权限 - 可以执行所有操作\n\n        ### 速率限制\n        - 匿名用户: 100 请求/分钟\n        - 认证用户: 200 请求/分钟\n        - 管理员: 500 请求/分钟\n\n        ### 默认账户\n        - 管理员: `admin` / `admin123`\n        - 普通用户: `user` / `user123`\n        ",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "开发环境"
    },
    {
      "url": "https://api.example.com",
      "description": "生产环境"
    }
  ],
  "paths": {
...
    "/api/v1/statistics/dashboard": {
      "get": {
        "tags": [
          "statistics"
        ],
        "summary": "Get Dashboard Stats",
        "description": "Get dashboard statistics.\nReturns key metrics for the main dashboard.",
        "operationId": "get_dashboard_stats_api_v1_statistics_dashboard_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/statistics/access-trends": {
      "get": {
        "tags": [
          "statistics"
        ],
        "summary": "Get Access Trends",
        "description": "Get access trends over time.\nReturns access statistics grouped by time periods.",
        "operationId": "get_access_trends_api_v1_statistics_access_trends_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "days",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 365,
              "minimum": 1,
              "description": "Number of days to analyze",
              "default": 30,
              "title": "Days"
            },
            "description": "Number of days to analyze"
          },
          {
            "name": "group_by",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^(hour|day|week)$",
              "description": "Grouping period",
              "default": "day",
              "title": "Group By"
            },
            "description": "Grouping period"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/statistics/department-stats": {
      "get": {
        "tags": [
          "statistics"
        ],
        "summary": "Get Department Statistics",
        "description": "Get statistics by department.\nReturns access statistics grouped by department.",
        "operationId": "get_department_statistics_api_v1_statistics_department_stats_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "start_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Start date for analysis",
              "title": "Start Date"
            },
            "description": "Start date for analysis"
          },
          {
            "name": "end_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "description": "End date for analysis",
              "title": "End Date"
            },
            "description": "End date for analysis"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/statistics/person-activity": {
      "get": {
        "tags": [
          "statistics"
        ],
        "summary": "Get Person Activity Stats",
        "description": "Get person activity statistics.\nReturns access statistics for individual persons.",
        "operationId": "get_person_activity_stats_api_v1_statistics_person_activity_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "start_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Start date for analysis",
              "title": "Start Date"
            },
            "description": "Start date for analysis"
          },
          {
            "name": "end_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "description": "End date for analysis",
              "title": "End Date"
            },
            "description": "End date for analysis"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "description": "Number of persons to return",
              "default": 50,
              "title": "Limit"
            },
            "description": "Number of persons to return"
          },
          {
            "name": "sort_by",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^(attempts|success_rate|last_access)$",
              "default": "attempts",
              "title": "Sort By"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/statistics/location-stats": {
      "get": {
        "tags": [
          "statistics"
        ],
        "summary": "Get Location Statistics",
        "description": "Get statistics by location.\nReturns access statistics grouped by location.",
        "operationId": "get_location_statistics_api_v1_statistics_location_stats_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "start_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Start date for analysis",
              "title": "Start Date"
            },
            "description": "Start date for analysis"
          },
          {
            "name": "end_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "description": "End date for analysis",
              "title": "End Date"
            },
            "description": "End date for analysis"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/statistics/recognition-performance": {
      "get": {
        "tags": [
          "statistics"
        ],
        "summary": "Get Recognition Performance Stats",
        "description": "Get face recognition performance statistics.\nReturns statistics about recognition accuracy and performance.",
        "operationId": "get_recognition_performance_stats_api_v1_statistics_recognition_performance_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "start_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Start date for analysis",
              "title": "Start Date"
            },
            "description": "Start date for analysis"
          },
          {
            "name": "end_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "description": "End date for analysis",
              "title": "End Date"
            },
            "description": "End date for analysis"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT认证令牌"
      }
    },
...
  },
  "security": [
    {
      "BearerAuth": []
    }
  ],
...
}